// client/src/api/types/websocket/WebsocketEventType.ts
export enum WebsocketEventType {
    MESSAGE_CREATED = "MESSAGE_CREATED",
    MESSAGE_UPDATED = "MESSAGE_UPDATED",
    MESSAGE_DELETED = "MESSAGE_DELETED",
    CHAT_UPDATED = "CHAT_UPDATED",
    USER_JOINED_CHAT = "USER_JOINED_CHAT",
    USER_LEFT_CHAT = "USER_LEFT_CHAT",
    USER_KICKED_FROM_CHAT = "USER_KICKED_FROM_CHAT",
    USER_BLOCKED_IN_CHAT = "USER_BLOCKED_IN_CHAT",
    USER_UPDATED = "USER_UPDATED",
    USER_DELETED = "USER_DELETED",
    CHAT_SUBSCRIPTION = "CHAT_SUBSCRIPTION",
    CHAT_UNSUBSCRIPTION = "CHAT_UNSUBSCRIPTION",
    MESSAGES_DELETED = "MESSAGES_DELETED",
    CHAT_BLOCKING_CREATED = "CHAT_BLOCKING_CREATED",
    CHAT_BLOCKING_UPDATED = "CHAT_BLOCKING_UPDATED",
    CHAT_BLOCKING_CANCELED = "CHAT_BLOCKING_CANCELED",
    CHAT_PARTICIPANT_WENT_ONLINE = "CHAT_PARTICIPANT_WENT_ONLINE",
    CHAT_PARTICIPANT_WENT_OFFLINE = "CHAT_PARTICIPANT_WENT_OFFLINE"
}